+++ package/core/src/components/mics/BottomSheet/DefaultProps.ts
@@ -1,0 +1,17 @@
+import { ColorTypeProps } from "../../../helpers/GetColorValue";
+
+export type BottomSheetStyleProps = {
+  backgroundColor: ColorTypeProps;
+  lineColor: ColorTypeProps;
+};
+
+export type BottomSheetComponentThemeProps = {
+  default: BottomSheetStyleProps;
+};
+
+export const BottomSheetComponentThemeData: BottomSheetComponentThemeProps = {
+  default: {
+    backgroundColor: "background",
+    lineColor: "muted",
+  },
+};