+++ package/core/src/components/typography/Heading/DefaultProps.ts
@@ -1,3 +1,4 @@
+import { TextStyle } from "react-native";
 import { ColorTypeProps } from "../../../helpers/GetColorValue";
 import { HeadingSizesProps } from "../../../types";
 
@@ -5,6 +6,17 @@ export type HeadingProps = {
   color?: ColorTypeProps;
   fontSize?: keyof HeadingSizesProps | number;
   fontFamily?: string;
+  textAlign?: TextStyle["textAlign"];
+  letterSpacing?: TextStyle["letterSpacing"];
+  lineHeight?: TextStyle["lineHeight"];
+  textDecorationLine?: TextStyle["textDecorationLine"];
+  textDecorationStyle?: TextStyle["textDecorationStyle"];
+  textDecorationColor?: TextStyle["textDecorationColor"];
+  textShadowColor?: TextStyle["textShadowColor"];
+  textShadowOffset?: TextStyle["textShadowOffset"];
+  textShadowRadius?: TextStyle["textShadowRadius"];
+  textTransform?: TextStyle["textTransform"];
+  testID?: TextStyle["testID"];
 };
 
 export type HeadingComponentThemeProps = {