+++ package/core/src/styles/Theme.ts
@@ -1,6 +1,8 @@
 import { ActionButtonComponentThemeData } from "../components/button/ActionButton/DefaultProps";
 import { ButtonComponentThemeData } from "../components/button/Button/DefaultProps";
 import { FormLayoutComponentThemeData } from "../components/form/_layout/DefaultProps";
+import { InputComponentThemeData } from "../components/form/_types/InputProps";
+import { BottomSheetComponentThemeData } from "../components/mics/BottomSheet/defaultProps";
 import { BoxComponentThemeData } from "../components/mics/Box/DefaultProps";
 import { GridComponentThemeData } from "../components/mics/Grid/DefaultProps";
 import { GroupComponentThemeData } from "../components/mics/Group/DefaultProps";
@@ -43,6 +45,8 @@ const Theme: ThemeInterface = {
     Button: ButtonComponentThemeData,
     ActionButton: ActionButtonComponentThemeData,
     FormLayout: FormLayoutComponentThemeData,
+    Input: InputComponentThemeData,
+    BottomSheet: BottomSheetComponentThemeData,
   },
   defaultOptions: {
     gap: 5,